Fix coverage with custom vocabulary
