+++ vitest.config.mjs
@@ -2,6 +2,7 @@ import { defineConfig } from 'vitest/config'
 
 export default defineConfig({
   test: {
+    globalSetup: ["tests/schema/oas-schema.mjs"],
     coverage: {
       provider: "custom",
       customProviderModule: "@hyperjump/json-schema-coverage/vitest/coverage-provider",
